# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/shipments": {
      "post": {
        "operationId": "ShipmentsController_createShipment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShipmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create shipment (Seller assigns 3PL)",
        "tags": [
          "Shipments"
        ]
      }
    },
    "/api/v1/shipments/me": {
      "get": {
        "operationId": "ShipmentsController_getMyShipments",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my shipments",
        "tags": [
          "Shipments"
        ]
      }
    },
    "/api/v1/shipments/{id}": {
      "get": {
        "operationId": "ShipmentsController_getShipmentById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get shipment by ID",
        "tags": [
          "Shipments"
        ]
      }
    },
    "/api/v1/shipments/{id}/milestones": {
      "post": {
        "operationId": "ShipmentsController_updateMilestone",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMilestoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update milestone (3PL, Seller)",
        "tags": [
          "Shipments"
        ]
      }
    },
    "/api/v1/shipments/{id}/documents": {
      "post": {
        "operationId": "ShipmentsController_uploadDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload document (Seller/3PL)",
        "tags": [
          "Shipments"
        ]
      }
    },
    "/api/v1/shipments/{id}/pod": {
      "post": {
        "operationId": "ShipmentsController_uploadPOD",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload POD (3PL)",
        "tags": [
          "Shipments"
        ]
      }
    },
    "/api/v1/shipments/{id}/documents/{docId}/verify": {
      "put": {
        "operationId": "ShipmentsController_verifyDocument",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Verify document (Admin)",
        "tags": [
          "Shipments"
        ]
      }
    },
    "/api/v1/shipments/{id}/tracking": {
      "get": {
        "operationId": "ShipmentsController_getTracking",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get shipment tracking",
        "tags": [
          "Shipments"
        ]
      }
    },
...
    "/api/v1/settlements/batches": {
      "post": {
        "operationId": "SettlementsController_createBatch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create settlement batch (Admin only)",
        "tags": [
          "Settlements"
        ]
      },
      "get": {
        "operationId": "SettlementsController_getAllBatches",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all settlement batches (Admin)",
        "tags": [
          "Settlements"
        ]
      }
    },
    "/api/v1/settlements/batches/{id}": {
      "get": {
        "operationId": "SettlementsController_getBatchById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get settlement batch by ID",
        "tags": [
          "Settlements"
        ]
      }
    },
    "/api/v1/settlements/batches/{id}/run-payouts": {
      "post": {
        "operationId": "SettlementsController_runPayouts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunPayoutsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Run payouts for batch (Admin)",
        "tags": [
          "Settlements"
        ]
      }
    },
    "/api/v1/settlements/payouts/me": {
      "get": {
        "operationId": "SettlementsController_getMyPayouts",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my payouts (Seller/3PL)",
        "tags": [
          "Settlements"
        ]
      }
    },
    "/api/v1/settlements/payouts/{id}": {
      "get": {
        "operationId": "SettlementsController_getPayoutById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get payout by ID",
        "tags": [
          "Settlements"
        ]
      }
    },
    "/api/v1/settlements/payouts/{id}/confirm": {
      "put": {
        "operationId": "SettlementsController_confirmPayout",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark payout as confirmed (Admin)",
        "tags": [
          "Settlements"
        ]
      }
    },
    "/api/v1/settlements/summary": {
      "get": {
        "operationId": "SettlementsController_getSettlementSummary",
        "parameters": [
          {
            "name": "startDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get settlement summary for period (Admin)",
        "tags": [
          "Settlements"
        ]
      }
    },
    "/api/v1/settlements/history/me": {
      "get": {
        "operationId": "SettlementsController_getMySettlementHistory",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get my settlement history (Seller/3PL)",
        "tags": [
          "Settlements"
        ]
      }
    },
...
    "/api/v1/admin/settlements/batches": {
      "post": {
        "operationId": "SettlementsAdminController_createBatch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Settlements"
        ]
      },
      "get": {
        "operationId": "SettlementsAdminController_listBatches",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Settlements"
        ]
      }
    },
    "/api/v1/admin/settlements/batches/{id}/run-payouts": {
      "post": {
        "operationId": "SettlementsAdminController_runPayouts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunPayoutsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Settlements"
        ]
      }
    },
...
  },
  "info": {
    "title": "B2B Marketplace API",
    "description": "Backend API for B2B Marketplace Platform",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateShipmentDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMilestoneDto": {
        "type": "object",
        "properties": {}
      },
      "UploadDocumentDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateBatchDto": {
        "type": "object",
        "properties": {}
      },
      "RunPayoutsDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}